+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 
 ## stream-chat-android-state
 ### 🐞 Fixed
+- Fix thread replies updating channel unread count. [#6056](https://github.com/GetStream/stream-chat-android/pull/6056)
 - Fix showing channels loading state when querying offline and no matching channels are found in the DB. [#6051](https://github.com/GetStream/stream-chat-android/pull/6051)
 
 ### ⬆️ Improved